requirements.txt
Django==6.0.2
djangorestframework
psycopg2-binary
django-filter
